LayoutTests/http/tests/dom/window-open-about-webkit-org-and-access-document-async-delegates-expected.txt

Tests opening a new about://webkit.org window and accessing its document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.name is "SecurityError"
PASS successfullyParsed is true

TEST COMPLETE
